pyproject.toml
[tool.poetry]
name = "simple-graph"
version = "0.1.0"
description = ""
authors = ["rxmrsd <morisada_keisuke@retail-ai.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.2.13"
langchain-community = "^0.2.12"
langchain-google-vertexai = "^1.0.8"
langchain-experimental = "^0.0.64"
langchain-text-splitters = "^0.2.2"
python-dotenv = "^1.0.1"
click = "^8.1.7"
neo4j = "^5.23.1"
langchain-openai = "^0.1.21"


[tool.poetry.group.dev.dependencies]
mypy = "^1.11.1"
ruff = "^0.5.7"
black = "^24.8.0"
ipykernel = "^6.29.5"
tiktoken = "^0.7.0"
json-repair = "^0.27.2"
wikipedia = "^1.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
show_column_numbers = true
ignore_missing_imports = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
warn_unused_ignores = true
warn_redundant_casts = true

[tool.black]
line-length = 79

[tool.ruff]
line-length = 79
select = [
    "E",   # pycodestype error
    "W",   # pycodestyle warning
    "F",   # pyflakes
    "C90", # mccabe
    "I",   # isort
    "D",   # pydocstyle
    "B",   # flake8-bugbear
    "COM", # flake8-commas
    "S",   # flake8-bandit 
    "PTH", # flake8-use-pathlib 
]
ignore = [
    "D415", # pydocstyle about First line should end with a '.' , '?' or '!'
    "E501", # line too long, handled by black
]
unfixable = [
    "F401", # disable auto-fix for unused-imports
    "F841", # disable auto-fix for unused-variable
]
target-version = "py311"

[tool.ruff.per-file-ignores]
# ignore assert error only in unnit-test
"test_*.py" = ["S101"]
"*_test.py" = ["S101"]

[tool.ruff.pydocstyle]
convention = "google"

[tool.ruff.mccabe]
max-complexity = 10